+++ tests/test_title_reference.py
@@ -147,6 +147,7 @@ def test_title_reference_parser_should_set_all_crew(imdb_id, n, crew):
 
 @pytest.mark.parametrize(("imdb_id", "episode_count"), [
     ("tt2121965", 176),  # House M.D.: Everybody Dies
+    ("tt1247466", 10),  # Band of Brothers: Points
 ])
 def test_title_parser_should_set_series_episode_count_for_episode(imdb_id, episode_count):
     parsed = web.get_title(imdb_id=imdb_id, page="reference")