+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 0.3 (unreleased)
 
+- Make TV mini-series compatible with series.
+
 ## 0.2 (unreleased)
 
 - New page parser: episodes.